move extensions to folders
