+++ Quiz/Application/TaskService/TaskService.cs
@@ -6,6 +6,7 @@ using Application.Repositories;
 using Domain.Entities;
 using Domain.Entities.TaskGenerators;
 using Domain.Values;
+using Infrastructure.Extensions;
 using Infrastructure.Result;
 using static Infrastructure.Result.None;
