+++ Quiz/Domain/Entities/TaskGenerators/TemplateTaskGenerator.cs
@@ -2,6 +2,7 @@ using System;
 using System.Linq;
 using Domain.Values;
 using Infrastructure;
+using Infrastructure.Extensions;
 using Scriban;
 using Scriban.Runtime;
 using static Infrastructure.Storage;
