+++ Quiz/Infrastructure/Extensions/EnumerableExtensions.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Infrastructure
+namespace Infrastructure.Extensions
 {
     public static class EnumerableExtensions
     {
