+++ Quiz/Infrastructure/Storage.cs
@@ -1,8 +1,9 @@
-﻿    using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics.Contracts;
 using System.Linq;
 using System.Text.RegularExpressions;
+using Infrastructure.Extensions;
 using static System.String;
 
 namespace Infrastructure
